+++ docs/src/advanced-usage/state-var-handler/index.js
@@ -127,7 +127,10 @@ export class StateVarHandler extends LitDocsContent(LitElement) {
 
             <h3>No-decorator usage</h3>
 
-            <p>Without decorators, you could implement it like this:</p>
+            <p>
+                Without decorators, use the static <code>stateVarOptions()</code>
+                getter method to specify the options:
+            </p>
 
             <p>
                 <code-block .code=${this.customStatevarNoDecoratorUsageCode}></code-block>
@@ -262,7 +265,7 @@ class DemoState extends LitState {
 
 class DemoState extends LitState {
 
-    static get stateVars() {
+    static get stateVarOptions() {
         return {
             myVar: localStorageStateVar({
                 localStorageKey: 'counter',